+++ tests/runtests.sh
@@ -5,6 +5,7 @@ set -ex
 export IS_BSST_TESTS_IS_IN_PROGRESS=1
 
 ./test_size_reports.sh
+./test_name_aliases.sh
 ./test_dynamic_stack_access.sh
 ./test_plugins.sh
 ./test_integer_conversion.py